+++ src/briefy/leica/models/job/assignment.py
@@ -160,11 +160,11 @@ class Assignment(AssignmentDates, mixins.AssignmentRolesMixin, mixins.Assignment
 
     _workflow = workflows.AssignmentWorkflow
 
-    __exclude_attributes__ = ['assets', ]
+    __exclude_attributes__ = ['assets', 'comments']
 
     __summary_attributes__ = __summary_attributes__
     __summary_attributes_relations__ = [
-        'project', 'comments', 'location', 'professional', 'customer',
+        'project', 'location', 'professional', 'customer',
         'pool'
     ]
     __listing_attributes__ = __listing_attributes__